+++ src/test/converter2/issues/gh2800.ts
@@ -1,0 +1,6 @@
+/**
+ * Module docs
+ * {@includeCode ./gh2800.ts}
+ * @module
+ */
+export const bug = true;